requirements.txt
streamlit==1.29.0
requests==2.31.0
python-dateutil==2.8.2
pandas==2.1.4
